Fixed the problem of trying to read a non existing temporary file.

+++ edu.illinois.codingspectator.saferecorder/src/edu/illinois/codingspectator/saferecorder/SafeRecorder.java
@@ -7,8 +7,8 @@ import java.io.File;
 import java.io.IOException;
 
 import edu.illinois.codingspectator.codingtracker.helpers.Debugger;
-import edu.illinois.codingspectator.codingtracker.helpers.ResourceHelper;
 import edu.illinois.codingspectator.codingtracker.helpers.Messages;
+import edu.illinois.codingspectator.codingtracker.helpers.ResourceHelper;
 import edu.illinois.codingspectator.data.CodingSpectatorDataPlugin;
 
 /**
@@ -49,9 +49,12 @@ public class SafeRecorder {
 		Debugger.debug("END COMMIT");
 		File tempFile= currentRecordFile;
 		currentRecordFile= mainRecordFile;
-		String tempContent= ResourceHelper.readFileContent(tempFile);
-		record(tempContent);
-		tempFile.delete();
+		//Temporary file does not exist at this point if nothing is recorded to it during data uploading
+		if (tempFile.exists()) {
+			String tempContent= ResourceHelper.readFileContent(tempFile);
+			record(tempContent);
+			tempFile.delete();
+		}
 	}
 
 	public synchronized void record(CharSequence text) {
